// src/@types/entities/Pokemon/Types/PokemonTypesEntity.types.ts
export enum EPokemonTypes {
  BUG = 'bug',
  DARK = 'dark',
  DRAGON = 'dragon',
  ELECTRIC = 'electric',
  FAIRY = 'fairy',
  FIRE = 'fire',
  FIGHTING = 'fighting',
  FLYING = 'flying',
  GRASS = 'grass',
  GHOST = 'ghost',
  GROUND = 'ground',
  ICE = 'ice',
  NORMAL = 'normal',
  POISON = 'poison',
  PSYCHIC = 'psychic',
  ROCK = 'rock',
  STEEL = 'steel',
  WATER = 'water',
}