// src/@types/entities/Stats/StatsEntity.types.ts
export enum EStatNames {
  HP = 'hp',
  ATTACK = 'attack',
  DEFENSE = 'defense',
  SPEED = 'speed',
  SPECIAL_ATTACK = 'special-attack',
  SPECIAL_DEFENSE = 'special-defense',
}

export enum EStatRanges {
  TERRIBLE = 'terrible',
  LOW = 'low',
  AVERAGE = 'average',
  GOOD = 'good',
  HIGH = 'high',
  EXCELLENT = 'excellent',
  LEGENDARY = 'legendary',
}